fix: address review — split test file, use dispatch with --verify
- Split HAO038-HAO045 tests into test_guard_remote_pair_proof.py
  (keeps test_guard_remote_pair_flow.py under 500 lines)
- Fix HAO040: use dispatch_guard_remote_pair_command with verify=True
  to actually exercise the --verify path including finalize_connect_payload
- Use OpenClaw runtime for HAO040 (Hermes always returns safe=True)
